split reducers, add utilities
